Move "Projects" section to its own page
The "Projects" section could've easily continued to live on the home
page, but by putting the "Projects" section in its own separate subpage,
it indicates that I'm not defined by the projects that I have done.
Obviously, at the moment it does, considering that's the only subpage I
have! But at least in the future I can easily have separate sub pages for
other things.

Technically, this unfortunately has resulted the duplication of the XML
file completely, and much of the XSL is duplicated.